Convert delta degrees from double to float for Unity. Showing radial controller rotation and clicking working in Demo scene.

+++ RadialControllerUnity/Assets/RadialController/Bridge/IRadialControllerPlatformBridge.cs
@@ -10,7 +10,7 @@ namespace RadialController {
          event Action onButtonPressed;
          event Action onButtonReleased;
          event Action onButtonHolding;
-         event Action<double> onRotationChanged;
+         event Action<float> onRotationChanged;
          event Action onControlAcquired;
          event Action onControlLost;
